Extract logic for associating GUIDs to collections
